+++ packages/contact-center/task/src/TaskList/index.tsx
@@ -1,12 +1,13 @@
 import React from 'react';
 import store from '@webex/cc-store';
 import {observer} from 'mobx-react-lite';
+import {ErrorBoundary} from 'react-error-boundary';
 
 import {TaskListComponent} from '@webex/cc-components';
 import {useTaskList} from '../helper';
 import {TaskListProps} from '../task.types';
 
-const TaskList: React.FunctionComponent<TaskListProps> = observer(
+const TaskListInternal: React.FunctionComponent<TaskListProps> = observer(
   ({onTaskAccepted, onTaskDeclined, onTaskSelected}) => {
     const {cc, taskList, currentTask, deviceType, logger} = store;
 
@@ -20,5 +21,18 @@ const TaskList: React.FunctionComponent<TaskListProps> = observer(
     return <TaskListComponent {...props} />;
   }
 );
+
+const TaskList: React.FunctionComponent<TaskListProps> = (props) => {
+  return (
+    <ErrorBoundary
+      fallbackRender={() => <></>}
+      onError={(error: Error) => {
+        if (store.onErrorCallback) store.onErrorCallback('TaskList', error);
+      }}
+    >
+      <TaskListInternal {...props} />
+    </ErrorBoundary>
+  );
+};
 
 export {TaskList};